+++ Albatross.Text.Table/release-notes.md
@@ -1,3 +1,7 @@
+# 8.0.5
+Add TextValue structure
+# 8.0.4
+Add StringTable.Column.NeverTruncate property
 # 8.0.2
 Add a filter function to [StringTable](./StringTable.cs)
 # 8.0.0